Massive upgrade
A massive cleanup + rework has been done to run on the latest netty available at the time (Netty 4.2.1 [May 2025]).
The project was converted to a gradle project.
Pathfinding has been reworked to be smarter.
Tickspeed has been adjusted, Comment and unused code cleanup
Combat rework, Skills rework, pet modifications and lots more.

+++ src/com/bestbudz/core/definitions/CombatSpellDefinition.java
@@ -9,64 +9,72 @@ import com.bestbudz.rs2.entity.item.Item;
 
 public class CombatSpellDefinition {
 
-	private String name;
-	private int id;
-	private byte baseMaxHit;
-	private double baseExperience;
-	private Animation animation;
-	private Graphic start;
-	private Projectile projectile;
-	private Graphic end;
-	private byte grade;
-	private int[] weapon;
-	private Item[] runes;
+  private String name;
+  private int id;
+  private byte baseMaxHit;
+  private double baseExperience;
+  private Animation animation;
+  private Graphic start;
+  private Projectile projectile;
+  private Graphic end;
+  private int grade;
+  private int[] weapon;
+  private Item[] runes;
 
-	public Animation getAnimation() {
-	return animation;
-	}
+  public Animation getAnimation() {
+    return animation;
+  }
 
-	public double getBaseExperience() {
-	return baseExperience;
-	}
+  public double getBaseExperience() {
+    return baseExperience;
+  }
 
-	public int getBaseMaxHit() {
-	return baseMaxHit;
-	}
+  public int getBaseMaxHit() {
+    return baseMaxHit;
+  }
 
-	public Graphic getEnd() {
-	return end;
-	}
+  public Graphic getEnd() {
+    return end;
+  }
 
-	public int getId() {
-	return id;
-	}
+  public int getId() {
+    return id;
+  }
 
-	public int getGrade() {
-	return grade;
-	}
+  public int getGrade() {
+    return grade;
+  }
 
-	public String getName() {
-	return name;
-	}
+  public String getName() {
+    return name;
+  }
 
-	public Projectile getProjectile() {
-	return projectile;
-	}
+  public Projectile getProjectile() {
+    return projectile;
+  }
 
-	public Item[] getRunes() {
-	return runes;
-	}
+  public Item[] getRunes() {
+    return runes;
+  }
 
-	public Graphic getStart() {
-	return start;
-	}
+  public Graphic getStart() {
+    return start;
+  }
 
-	public int[] getWeapons() {
-	return weapon;
-	}
+  public int[] getWeapons() {
+    return weapon;
+  }
 
-	@Override
-	public String toString() {
-	return "CombatSpellDefinition [name=" + name + ", id=" + id + ", grade=" + grade + ", runes=" + Arrays.toString(runes) + "]";
-	}
+  @Override
+  public String toString() {
+    return "CombatSpellDefinition [name="
+        + name
+        + ", id="
+        + id
+        + ", grade="
+        + grade
+        + ", runes="
+        + Arrays.toString(runes)
+        + "]";
+  }
 }
