+++ src/com/bestbudz/rs2/content/achievements/AchievementButtons.java
@@ -1,16 +1,9 @@
 package com.bestbudz.rs2.content.achievements;
 
-import java.util.HashMap;
-
 import com.bestbudz.rs2.content.achievements.AchievementHandler.AchievementDifficulty;
 import com.bestbudz.rs2.entity.stoner.Stoner;
+import java.util.HashMap;
 
-/**
- * Handles the achievement buttons
- * 
- * @author Jaybane
- * @author Jaybane
- */
 public class AchievementButtons {
 
 	private static final HashMap<Integer, AchievementList> BUTTONS = new HashMap<Integer, AchievementList>();
@@ -28,8 +21,8 @@ public class AchievementButtons {
 	}
 
 	public static boolean handleButtons(Stoner stoner, int buttonId) {
-	if (BUTTONS.containsKey((Integer) buttonId)) {
-		AchievementInterface.sendInterfaceForAchievement(stoner, BUTTONS.get((Integer) buttonId));
+	if (BUTTONS.containsKey(buttonId)) {
+		AchievementInterface.sendInterfaceForAchievement(stoner, BUTTONS.get(buttonId));
 		return true;
 	}
 	return false;
